+++ assets/r/transform_clean_delete_selected.R
@@ -3,3 +3,6 @@
 # Remove specific variables from the dataset.
 
 ds$SELECTED_VAR <- NULL
+
+glimpse(ds)
+summary(ds)