+++ lib/r/remove_windows_chars.dart
@@ -1,3 +1,0 @@
-String removeWindowsChars(String txt) {
-  return txt.replaceAll(RegExp(r'\[\d{1,2}C'), '');
-}