+++ lib/tabs/explore.dart
@@ -5,7 +5,7 @@
 /// License: https://www.gnu.org/licenses/gpl-3.0.en.html
 ///
 //
-// Time-stamp: <Wednesday 2024-07-31 08:38:31 +1000 Graham Williams>
+// Time-stamp: <Wednesday 2024-08-14 16:23:16 +1000 Graham Williams>
 //
 // Licensed under the GNU General Public License, Version 3 (the "License");
 //
@@ -37,7 +37,6 @@ import 'package:rattle/features/missing/panel.dart';
 import 'package:rattle/features/correlation/panel.dart';
 import 'package:rattle/features/tests/panel.dart';
 import 'package:rattle/features/interactive/panel.dart';
-//TODO import 'package:rattle/features/wordcloud/panel.dart';
 import 'package:rattle/providers/explore.dart';
 
 final List<Map<String, dynamic>> explorePanels = [
@@ -107,7 +106,7 @@ class _ExploreTabsState extends ConsumerState<ExploreTabs>
   @override
   Widget build(BuildContext context) {
     super.build(context);
-    debugPrint('EXPLORE: Rebuild.');
+    debugPrint('REBUILDING\t\tExploreTab.');
 
     return Column(
       children: [