+++ lib/tabs/transform.dart
@@ -5,7 +5,7 @@
 /// License: https://www.gnu.org/licenses/gpl-3.0.en.html
 ///
 //
-// Time-stamp: <Friday 2024-08-02 13:40:38 +1000 Graham Williams>
+// Time-stamp: <Wednesday 2024-08-14 16:23:43 +1000 Graham Williams>
 //
 // Licensed under the GNU General Public License, Version 3 (the "License");
 //
@@ -91,7 +91,7 @@ class _TransformTabsState extends ConsumerState<TransformTabs>
   @override
   Widget build(BuildContext context) {
     super.build(context);
-    debugPrint('TRANSFORM: Rebuild.');
+    debugPrint('REBUILDING\t\tTransformTab.');
 
     return Column(
       children: [