+++ lib/widgets/text_page.dart
@@ -1,6 +1,6 @@
 /// Helper widget to build the common text based pages.
 //
-// Time-stamp: <Thursday 2024-07-25 12:14:15 +1000 Graham Williams>
+// Time-stamp: <Monday 2024-08-12 08:11:10 +1000 Graham Williams>
 //
 /// Copyright (C) 2024, Togaware Pty Ltd
 ///
@@ -26,13 +26,12 @@
 library;
 
 // Group imports by dart, flutter, packages, local. Then alphabetically.
-
 import 'package:flutter/material.dart';
-
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:rattle/constants/app.dart';
+import 'package:rattle/constants/spacing.dart';
 import 'package:rattle/constants/sunken_box_decoration.dart';
 import 'package:rattle/utils/word_wrap.dart';
 
@@ -64,9 +63,28 @@ class TextPage extends StatelessWidget {
                 launchUrl(url);
               },
             ),
-            SelectableText(
-              content,
-              style: monoTextStyle,
+            // Wrap SelectableText in SingleChildScrollView for horizontal scrolling.
+
+            SingleChildScrollView(
+              scrollDirection: Axis.horizontal,
+              child: SelectableText(
+                content,
+                style: monoTextStyle,
+                // Ensure that text does not wrap, allowing horizontal scroll.
+
+                textAlign: TextAlign.left,
+              ),
+            ),
+            // 20240812 gjw Add a bottom spacer to leave a gap for the page
+            // navigation whenscrolling to the bottom of the page so that it can
+            // be visible in at least some part of any very busy pages.
+            textPageBottomSpace,
+            // 20240812 gjw Add a divider to mark the end of the text page.
+            const Divider(
+              thickness: 15,
+              color: Color(0XFFBBDEFB),
+              indent: 0,
+              endIndent: 20,
             ),
           ],
         ),